chore(workers): remove unreachable code
