+++ projects/workers/src/worker/pipes/worker.pipe.ts
@@ -1,7 +1,6 @@
 import {Pipe, PipeTransform} from '@angular/core';
 import {Observable} from 'rxjs';
 import {WebWorker} from '../classes/web-worker';
-import {WorkerExecutor} from '../services/worker-executor.service';
 import {WorkerFunction} from '../types/worker-function';
 
 @Pipe({
@@ -10,10 +9,8 @@ import {WorkerFunction} from '../types/worker-function';
 export class WorkerPipe implements PipeTransform {
     private workers = new WeakMap<WorkerFunction, WebWorker>();
 
-    constructor(private workerExecutor: WorkerExecutor) {}
-
     transform<T, R>(value: T, fn: WorkerFunction<T, R>): Observable<R> {
-        const worker = this.workers.get(fn) || this.workerExecutor.createWorker(fn);
+        const worker = this.workers.get(fn) || WebWorker.fromFunction(fn);
 
         this.workers.set(fn, worker);
