+++ projects/workers/src/worker/worker.module.ts
@@ -1,12 +1,8 @@
-import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {WorkerPipe} from './pipes/worker.pipe';
-import {WorkerExecutor} from './services/worker-executor.service';
 
 @NgModule({
-    imports: [CommonModule],
     declarations: [WorkerPipe],
     exports: [WorkerPipe],
-    providers: [WorkerExecutor],
 })
 export class WorkerModule {}
